Breaks tests, enables deploy
I'm still getting to grips with how Chalice likes to be tested vs how it deploys. The upshot of this is that, right now, I can have tests or I can have deployment but not both. I'm going to investigate how to manage this more effectively and slightly more in line with how I'd do it in Flask